% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_api.R
\name{fhir_api}
\alias{fhir_api}
\title{Send a GET request to a FHIR API}
\usage{
fhir_api(
  path,
  base_url = "http://hapi.fhir.org/baseR4",
  delay = 5,
  user_agent,
  ...
)
}
\arguments{
\item{path}{Path from the base url.}

\item{base_url}{Base url.}

\item{delay}{Delay in seconds before making the API call.}

\item{user_agent}{(Optional) User agent.}
}
\value{
A fhir_test_api class object.
}
\description{
Send a GET request to a FHIR API
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{http_type}},\code{\link[httr]{content}},\code{\link[httr]{http_error}},\code{\link[httr]{status_code}}
\code{\link[jsonlite]{toJSON, fromJSON}}
}
