% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-fhir_test_api.R
\name{fhir_test_api}
\alias{fhir_test_api}
\title{Send a GET request to the FHIR Test API}
\usage{
fhir_test_api(path, user_agent)
}
\arguments{
\item{path}{Path from \url{http://test.fhir.org/r4/} base url.}

\item{user_agent}{(Optional) User agent.}
}
\value{
A fhir_test_api class object.
}
\description{
Send a GET request to the FHIR Test API
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{http_type}},\code{\link[httr]{content}},\code{\link[httr]{http_error}},\code{\link[httr]{status_code}}
\code{\link[jsonlite]{toJSON, fromJSON}}
}
