% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FHIRworks-package.R
\docType{package}
\name{FHIRworks-package}
\alias{FHIRworks}
\alias{FHIRworks-package}
\title{FHIRworks: What the Package Does (One Line, Title Case)}
\description{
FHIRworks is an R interface for making FHIR API calls.
Learn more about FHIR at \url{http://build.fhir.org/http.html}.
}
\author{
\strong{Maintainer}: Meera Patel, MD \email{patelmeeray@gmail.com} (\href{https://orcid.org/0000-0002-3401-3948}{ORCID})

}
\keyword{internal}
