% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-get_resource_entity_set.R
\name{get_resource_entity_set}
\alias{get_resource_entity_set}
\title{Get a Resource's Entity Set}
\usage{
get_resource_entity_set(resource, user_agent)
}
\arguments{
\item{resource}{FHIR resource.}

\item{user_agent}{(Optional) User agent.}
}
\value{
A fhir_test_api class object.
}
\description{
Get a Resource's Entity Set
}
